Move structs for TopicFrame and FragmentFrame to coordinator-helper package
