+++ common/fragment.go
@@ -1,7 +1,6 @@
 package common
 
 import (
-	"encoding/binary"
 	"math/rand"
 )
 
@@ -9,43 +8,4 @@ const MaxFragmentCount = 1 << 8
 
 func GenerateFragmentId() uint32 {
 	return uint32(rand.Intn(MaxFragmentCount)) + 1
-}
-
-type FrameForFragment struct {
-	data []byte
-}
-
-func NewFrameForFragment(data []byte) *FrameForFragment {
-	return &FrameForFragment{data: data}
-}
-
-func NewFrameForFragmentFromValues(lastOffset uint64, numSubscribers uint64) *FrameForFragment {
-	data := make([]byte, uint64Len*2)
-	binary.BigEndian.PutUint64(data[0:], lastOffset)
-	binary.BigEndian.PutUint64(data[uint64Len:uint64Len*2], numSubscribers)
-	return &FrameForFragment{data: data}
-}
-
-func (f FrameForFragment) Data() []byte {
-	return f.data
-}
-
-func (f FrameForFragment) Size() int {
-	return len(f.data)
-}
-
-func (f FrameForFragment) LastOffset() uint64 {
-	return binary.BigEndian.Uint64(f.Data()[:uint64Len])
-}
-
-func (f *FrameForFragment) SetLastOffset(offset uint64) {
-	binary.BigEndian.PutUint64(f.data[0:uint64Len], offset)
-}
-
-func (f FrameForFragment) NumSubscribers() uint64 {
-	return binary.BigEndian.Uint64(f.Data()[uint64Len : uint64Len*2])
-}
-
-func (f *FrameForFragment) SetNumSubscribers(num uint64) {
-	binary.BigEndian.PutUint64(f.data[uint64Len:uint64Len*2], num)
 }
